refactor mock logwriters
