+++ aws-shared/src/test/java/com/kdgregory/logging/testhelpers/cloudwatch/MockCloudWatchWriter.java
@@ -14,57 +14,21 @@
 
 package com.kdgregory.logging.testhelpers.cloudwatch;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import com.kdgregory.logging.aws.cloudwatch.CloudWatchConstants;
 import com.kdgregory.logging.aws.cloudwatch.CloudWatchWriterConfig;
 import com.kdgregory.logging.common.LogMessage;
-import com.kdgregory.logging.common.LogWriter;
-import com.kdgregory.logging.common.util.DiscardAction;
+import com.kdgregory.logging.testhelpers.MockLogWriter;
 
 
 /**
  *  A mock equivalent of CloudWatchLogWriter, used by appender tests.
  */
 public class MockCloudWatchWriter
-implements LogWriter
+extends MockLogWriter<CloudWatchWriterConfig>
 {
-    public CloudWatchWriterConfig config;
-
-    public List<LogMessage> messages = new ArrayList<LogMessage>();
-    public LogMessage lastMessage;
-
-    public boolean stopped;
-
-
     public MockCloudWatchWriter(CloudWatchWriterConfig config)
     {
-        this.config = config;
-    }
-
-//----------------------------------------------------------------------------
-//  LogWriter
-//----------------------------------------------------------------------------
-
-    @Override
-    public void setBatchDelay(long value)
-    {
-        this.config.batchDelay = value;
-    }
-
-
-    @Override
-    public void setDiscardThreshold(int value)
-    {
-        this.config.discardThreshold = value;
-    }
-
-
-    @Override
-    public void setDiscardAction(DiscardAction value)
-    {
-        this.config.discardAction = value;
+        super(config);
     }
 
 
@@ -73,71 +37,5 @@ implements LogWriter
     {
         // this is copied from CloudWatchLogWriter
         return (message.size() + CloudWatchConstants.MESSAGE_OVERHEAD)  >= CloudWatchConstants.MAX_BATCH_BYTES;
-    }
-
-
-    @Override
-    public void addMessage(LogMessage message)
-    {
-        messages.add(message);
-        lastMessage = message;
-    }
-
-
-    @Override
-    public boolean initialize()
-    {
-        return true;
-    }
-
-
-    @Override
-    public boolean waitUntilInitialized(long millisToWait)
-    {
-        return true;
-    }
-
-
-    @Override
-    public void processBatch()
-    {
-        // nothing happening here
-    }
-
-
-    @Override
-    public void stop()
-    {
-        stopped = true;
-    }
-
-
-    @Override
-    public void shutdown()
-    {
-        // nothing happening here
-    }
-
-//----------------------------------------------------------------------------
-//  Runnable
-//----------------------------------------------------------------------------
-
-    @Override
-    public void run()
-    {
-        // we're not expecting to be on a background thread, so do nothing
-    }
-
-//----------------------------------------------------------------------------
-//  Mock-specific methods
-//----------------------------------------------------------------------------
-
-    /**
-     *  Returns the text for the numbered message (starting at 0).
-     */
-    public String getMessage(int msgnum)
-    throws Exception
-    {
-        return messages.get(msgnum).getMessage();
     }
 }
