+++ aws-shared/src/test/java/com/kdgregory/logging/testhelpers/sns/MockSNSWriter.java
@@ -14,37 +14,22 @@
 
 package com.kdgregory.logging.testhelpers.sns;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import com.kdgregory.logging.aws.sns.SNSWriterConfig;
 import com.kdgregory.logging.common.LogMessage;
-import com.kdgregory.logging.common.LogWriter;
-import com.kdgregory.logging.common.util.DiscardAction;
+import com.kdgregory.logging.testhelpers.MockLogWriter;
 
 
 /**
  *  A mock equivalent of SNSLogWriter, used by appender tests.
  */
 public class MockSNSWriter
-implements LogWriter
+extends MockLogWriter<SNSWriterConfig>
 {
-    public SNSWriterConfig config;
-
-    public List<LogMessage> messages = new ArrayList<LogMessage>();
-    public LogMessage lastMessage;
-
-    public boolean stopped;
-
-
     public MockSNSWriter(SNSWriterConfig config)
     {
-        this.config = config;
+        super(config);
     }
 
-//----------------------------------------------------------------------------
-//  LogWriter
-//----------------------------------------------------------------------------
 
     @Override
     public void setBatchDelay(long value)
@@ -54,88 +39,9 @@ implements LogWriter
 
 
     @Override
-    public void setDiscardThreshold(int value)
-    {
-        // ignored for now
-    }
-
-    @Override
-    public void setDiscardAction(DiscardAction value)
-    {
-        // ignored for now
-    }
-
-
-    @Override
     public boolean isMessageTooLarge(LogMessage message)
     {
         // there are no tests for this, so we'll pretend everything's great
         return false;
-    }
-
-
-    @Override
-    public void addMessage(LogMessage message)
-    {
-        messages.add(message);
-        lastMessage = message;
-    }
-
-
-    @Override
-    public boolean initialize()
-    {
-        return true;
-    }
-
-
-    @Override
-    public boolean waitUntilInitialized(long millisToWait)
-    {
-        return true;
-    }
-
-
-    @Override
-    public void processBatch()
-    {
-        // nothing happening here
-    }
-
-
-    @Override
-    public void stop()
-    {
-        stopped = true;
-    }
-
-
-    @Override
-    public void shutdown()
-    {
-        // nothing happening here
-    }
-
-//----------------------------------------------------------------------------
-//  Runnable
-//----------------------------------------------------------------------------
-
-    @Override
-    public void run()
-    {
-        // we're not expecting to be on a background thread, so do nothing
-    }
-
-//----------------------------------------------------------------------------
-//  Mock-specific methods
-//----------------------------------------------------------------------------
-
-    /**
-     *  Returns the text for the numbered message (starting at 0).
-     */
-    public String getMessage(int msgnum)
-    throws Exception
-    {
-        return messages.get(msgnum).getMessage();
     }
 }
